Move Identity and Session to types
