+++ portal/src/graphql/adminapi/UserDetailCommandBarContainer.tsx
@@ -3,7 +3,8 @@ import { CommandButton, ICommandBarItemProps } from "@fluentui/react";
 import { Context } from "@oursky/react-messageformat";
 import { useSystemConfig } from "../../context/SystemConfigContext";
 import SetUserDisabledDialog from "./SetUserDisabledDialog";
-import { extractUserInfoFromIdentities, Identity } from "../../util/user";
+import { extractUserInfoFromIdentities } from "../../util/user";
+import { Identity } from "../../types";
 import DeleteUserDialog from "./DeleteUserDialog";
 import { useNavigate } from "react-router-dom";
 import CommandBarContainer from "../../CommandBarContainer";
